feat: :sparkles: Introduce feature create a new lesson
This feature propose at the user to create a new lesson. It proposes to choose a date and a lesson type first, then the system checks if there is an accredited instructor able to give this type of lesson, then it checks if there is an available instructor. If no it displays an error message, if yes the lesson is created and a success message is displayed