feat: :sparkles: Skiers can book for an upcoming lesson they didn't already have booked.
Add a button that is available after having selected a skier. After having clicked on that button a Frame opens with all the upcoming lessons the skiers haven't already booked. The user has to selected one to add the booking. A checkbox that proposes to take an insurance out is also available. Then, either a success message is displayed if the booking has been added into the database or an error message saying that an error ocurrend while processing.